Split ColumnsValue tests into own module
